requirements.txt
djangorestframework
Django==5.0.6
djangorestframework==3.13.1
reportlab==3.6.7
chardet==5.0.0
# Ajoutez ici d'autres dépendances nécessaires
